youtube_summariser_backend/youtube_summariser_requirements.txt
uvicorn
fastapi

youtube-transcript-api
youtube_dl
pytube
openai-whisper

sentencepiece
transformers

nltk
scikit-learn